Text Carre Noir > GillesXYZ, Script WaitForSecondsTextChangeColor with waitDuration cast to float. In this way, we can be more prcise in the waiting duration before changing the color of a text

+++ Assets/Scripts/WaitForSecondsTextChangeColor.cs
@@ -6,7 +6,7 @@ public class WaitForSecondsTextChangeColor : MonoBehaviour {
 
     public Text displayedText;
 
-    public int waitDuration = 5;
+    public float waitDuration = 5;
 
 
     public byte displayedTextColorR0 = 255;
